zip: Refactor Zip operations into ZipActionImpl
ContainerActionImpl is more general than zip operations.
