+++ org.eclipse.transformer/src/main/java/org/eclipse/transformer/action/impl/EarActionImpl.java
@@ -17,7 +17,7 @@ import org.eclipse.transformer.action.SelectionRule;
 import org.eclipse.transformer.action.SignatureRule;
 import org.slf4j.Logger;
 
-public class EarActionImpl extends ContainerActionImpl {
+public class EarActionImpl extends ZipActionImpl {
 	public EarActionImpl(Logger logger, InputBuffer buffer,
 		SelectionRule selectionRule, SignatureRule signatureRule) {
 		super(logger, buffer, selectionRule, signatureRule);
