+++ org.eclipse.transformer/src/main/java/org/eclipse/transformer/action/impl/WarActionImpl.java
@@ -17,7 +17,7 @@ import org.eclipse.transformer.action.SelectionRule;
 import org.eclipse.transformer.action.SignatureRule;
 import org.slf4j.Logger;
 
-public class WarActionImpl extends ContainerActionImpl {
+public class WarActionImpl extends ZipActionImpl {
 	public WarActionImpl(Logger logger, InputBuffer buffer, SelectionRule selectionRule, SignatureRule signatureRule) {
 		super(logger, buffer, selectionRule, signatureRule);
 	}
